app completed

+++ lib/widgets/cart_item.dart
@@ -14,9 +14,73 @@ class CartItem extends StatefulWidget {
 
 class _CartItemState extends State<CartItem> {
   void removeFromCart() {
-    Provider.of<CartProvider>(context, listen: false)
-        .deleteProduct(widget.cartItem);
-    customSnackBar(context, darkGreen, "Product removed successfully");
+    showDialog(
+      context: context,
+      builder: (context) => AlertDialog(
+        titlePadding: const EdgeInsets.all(20),
+        contentPadding: const EdgeInsets.symmetric(horizontal: 24),
+        actionsPadding: const EdgeInsets.all(16),
+        shape: RoundedRectangleBorder(
+          borderRadius: BorderRadius.circular(
+            18,
+          ),
+        ),
+        title: const Row(
+          children: [
+            Icon(
+              Icons.info_rounded,
+              weight: 4,
+            ),
+            SizedBox(
+              width: 12,
+            ),
+            Text(
+              "Remove Product",
+              style: TextStyle(
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ],
+        ),
+        content: const Text(
+          "Are you sure to remove this product from cart?",
+          style: TextStyle(
+            fontSize: 16,
+          ),
+        ),
+        actions: [
+          TextButton(
+            onPressed: () {
+              Navigator.of(context).pop();
+            },
+            child: const Text(
+              "No",
+              style: TextStyle(
+                color: Colors.blue,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ),
+          TextButton(
+            onPressed: () {
+              Provider.of<CartProvider>(context, listen: false)
+                  .deleteProduct(widget.cartItem);
+              Navigator.of(context).pop();
+              customSnackBar(context, CustomColors.darkGreen,
+                  "Product removed successfully");
+            },
+            child: const Text(
+              "Yes",
+              style: TextStyle(
+                color: Colors.red,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+          ),
+        ],
+      ),
+      barrierDismissible: false,
+    );
   }
 
   @override
@@ -49,7 +113,7 @@ class _CartItemState extends State<CartItem> {
             },
             child: const Icon(
               Icons.delete,
-              color: darkRed,
+              color: CustomColors.darkRed,
             ),
           ),
         ),
